+++ src/topology.test.ts
@@ -4,13 +4,11 @@ import {
   getInputData,
   filterDAG,
   getNodesReadyToRun,
-  initResources,
   initSnapshotData,
   runTopology,
   resumeTopology,
   getResumeSnapshot,
   TopologyError,
-  cleanupResources,
 } from './topology'
 import { DAG, RunFn, Snapshot, Spec } from './types'
 import timers from 'timers/promises'
@@ -23,42 +21,27 @@ const dag: DAG = {
 }
 
 const spec: Spec = {
-  resources: {
-    elasticCloud: {
-      init: async () => 'elastic',
-    },
-    mongoDb: {
-      init: async () => 'mongo',
-    },
-    config: {
-      init() {
-        return { host: 'localhost' }
-      },
+  api: {
+    deps: [],
+    run: async () => [1, 2, 3],
+  },
+  details: {
+    deps: ['api'],
+    run: async ({ data }) => {
+      const ids: number[] = data[0]
+      return ids.reduce((acc, n) => _.set(n, `description ${n}`, acc), {})
     },
   },
-  nodes: {
-    api: {
-      run: async () => [1, 2, 3],
-      resources: ['elasticCloud'],
+  attachments: {
+    deps: ['api'],
+    run: async ({ data }) => {
+      const ids: number[] = data[0]
+      return ids.reduce((acc, n) => _.set(n, `file${n}.jpg`, acc), {})
     },
-    details: {
-      run: async ({ data }) => {
-        const ids: number[] = data[0]
-        return ids.reduce((acc, n) => _.set(n, `description ${n}`, acc), {})
-      },
-      resources: ['elasticCloud', 'mongo'],
-    },
-    attachments: {
-      run: async ({ data }) => {
-        const ids: number[] = data[0]
-        return ids.reduce((acc, n) => _.set(n, `file${n}.jpg`, acc), {})
-      },
-      resources: [],
-    },
-    writeToDB: {
-      run: async () => null,
-      resources: ['mongo', 'config'],
-    },
+  },
+  writeToDB: {
+    deps: ['details', 'attachments'],
+    run: async () => null,
   },
 }
 
@@ -91,6 +74,7 @@ describe('getNodesReadyToRun', () => {
   test('resume - pending', () => {
     const nodes = getNodesReadyToRun(dag, {
       api: {
+        deps: [],
         status: 'pending',
         started: new Date('2022-01-01T12:00:00Z'),
         input: {
@@ -104,6 +88,7 @@ describe('getNodesReadyToRun', () => {
   test('deps met - exclude completed', () => {
     const nodes = getNodesReadyToRun(dag, {
       api: {
+        deps: [],
         status: 'completed',
         started: new Date('2022-01-01T12:00:00Z'),
         finished: new Date('2022-01-01T12:05:00Z'),
@@ -119,6 +104,7 @@ describe('getNodesReadyToRun', () => {
   test('deps met - exclude completed and running', () => {
     const nodes = getNodesReadyToRun(dag, {
       api: {
+        deps: [],
         status: 'completed',
         started: new Date('2022-01-01T12:00:00Z'),
         finished: new Date('2022-01-01T12:05:00Z'),
@@ -129,6 +115,7 @@ describe('getNodesReadyToRun', () => {
         output: ['123', '456'],
       },
       details: {
+        deps: ['api'],
         status: 'running',
         started: new Date('2022-01-01T12:00:00Z'),
         input: [['123', '456']],
@@ -139,6 +126,7 @@ describe('getNodesReadyToRun', () => {
   test('deps met - exclude completed and errored', () => {
     const nodes = getNodesReadyToRun(dag, {
       api: {
+        deps: [],
         status: 'completed',
         started: new Date('2022-01-01T12:00:00Z'),
         finished: new Date('2022-01-01T12:05:00Z'),
@@ -149,6 +137,7 @@ describe('getNodesReadyToRun', () => {
         output: ['123', '456'],
       },
       details: {
+        deps: ['api'],
         status: 'errored',
         started: new Date('2022-01-01T12:00:00Z'),
         input: [['123', '456']],
@@ -159,6 +148,7 @@ describe('getNodesReadyToRun', () => {
   test('deps not met', () => {
     const nodes = getNodesReadyToRun(dag, {
       api: {
+        deps: [],
         status: 'completed',
         started: new Date('2022-01-01T12:00:00Z'),
         finished: new Date('2022-01-01T12:05:00Z'),
@@ -169,12 +159,14 @@ describe('getNodesReadyToRun', () => {
         output: ['123', '456'],
       },
       details: {
+        deps: ['api'],
         status: 'completed',
         started: new Date('2022-01-01T12:00:00Z'),
         input: [['123', '456']],
         output: { '123': 'foo', '456': 'bar' },
       },
       attachments: {
+        deps: ['api'],
         status: 'errored',
         started: new Date('2022-01-01T12:00:00Z'),
         input: [['123', '456']],
@@ -188,58 +180,11 @@ describe('getNodesReadyToRun', () => {
   })
 })
 
-describe('initResources', () => {
-  test('resources initialized', async () => {
-    expect(
-      await initResources(spec, ['elasticCloud', 'mongoDb', 'config'])
-    ).toEqual({
-      elasticCloud: 'elastic',
-      mongoDb: 'mongo',
-      config: { host: 'localhost' },
-    })
-  })
-})
-
-describe('cleanupResources', () => {
-  test('clean up initialized resources', async () => {
-    const cleanedUp: any[] = []
-    const cleanup = (x: any) => {
-      cleanedUp.push(x)
-    }
-
-    const spec: Spec = {
-      resources: {
-        elasticCloud: {
-          init: async () => 'elastic',
-          cleanup: async (x: any) => cleanup(x),
-        },
-        mongoDb: {
-          init: async () => 'mongo',
-          cleanup: (x: any) => cleanup(x),
-        },
-        config: {
-          init() {
-            return { host: 'localhost' }
-          },
-        },
-      },
-      nodes: {},
-    }
-
-    const initialized = await initResources(spec, [
-      'elasticCloud',
-      'mongoDb',
-      'config',
-    ])
-    await cleanupResources(spec, initialized)
-    expect(cleanedUp).toEqual(['elastic', 'mongo'])
-  })
-})
-
 describe('getInputData', () => {
   test('single dep', () => {
     const input = getInputData(dag, 'details', {
       api: {
+        deps: [],
         status: 'completed',
         started: new Date('2022-01-01T12:00:00Z'),
         finished: new Date('2022-01-01T12:05:00Z'),
@@ -255,6 +200,7 @@ describe('getInputData', () => {
   test('multiple deps', () => {
     const input = getInputData(dag, 'writeToDB', {
       api: {
+        deps: [],
         status: 'completed',
         started: new Date('2022-01-01T12:00:00Z'),
         finished: new Date('2022-01-01T12:05:00Z'),
@@ -265,6 +211,7 @@ describe('getInputData', () => {
         output: ['123', '456'],
       },
       details: {
+        deps: ['api'],
         status: 'completed',
         started: new Date('2022-01-01T12:00:00Z'),
         finished: new Date('2022-01-01T12:05:00Z'),
@@ -272,6 +219,7 @@ describe('getInputData', () => {
         output: { 123: { description: 'foo' } },
       },
       attachments: {
+        deps: ['api'],
         status: 'completed',
         started: new Date('2022-01-01T12:00:00Z'),
         finished: new Date('2022-01-01T12:05:00Z'),
@@ -287,6 +235,7 @@ describe('getInputData', () => {
   test('resume scenario', () => {
     const input = getInputData(dag, 'api', {
       api: {
+        deps: [],
         status: 'pending',
         input: {
           startDate: '2020-01-01',
@@ -305,70 +254,59 @@ describe('getInputData', () => {
 describe('initData', () => {
   test('data passed', () => {
     expect(initSnapshotData(dag, [1, 2, 3])).toEqual({
-      api: { input: [1, 2, 3] },
+      api: { deps: [], status: 'pending', input: [[1, 2, 3]] },
+      details: { deps: ['api'], status: 'pending' },
+      attachments: { deps: ['api'], status: 'pending' },
+      writeToDB: { deps: ['details', 'attachments'], status: 'pending' },
     })
   })
   test('data empty', () => {
-    expect(initSnapshotData(dag, {})).toEqual({})
+    expect(initSnapshotData(dag)).toEqual({
+      api: { deps: [], status: 'pending' },
+      details: { deps: ['api'], status: 'pending' },
+      attachments: { deps: ['api'], status: 'pending' },
+      writeToDB: { deps: ['details', 'attachments'], status: 'pending' },
+    })
   })
 })
 
 describe('runTopology', () => {
   test('nodes receive expected input', async () => {
-    const dag = {
-      api: { deps: [] },
-      details: { deps: ['api'] },
-    }
     const spec: Spec = {
-      resources: {
-        elasticCloud: {
-          init: async () => 'elastic',
-        },
-        mongoDb: {
-          init: async () => 'mongo',
-        },
+      api: {
+        deps: [],
+        run: async ({ data, context }) => ({
+          data,
+          context,
+        }),
       },
-      nodes: {
-        api: {
-          run: async ({ data, resources, context }) => ({
-            data,
-            resources,
-            context,
-          }),
-          resources: ['elasticCloud'],
-        },
-        details: {
-          run: async ({ data, resources, context }) => ({
-            data,
-            resources,
-            context,
-          }),
-          resources: ['mongoDb'],
-        },
+      details: {
+        deps: ['api'],
+        run: async ({ data, context }) => ({
+          data,
+          context,
+        }),
       },
     }
     const data = [1, 2, 3]
     const context = { launchMissleCode: 1234 }
-    const { promise, getSnapshot } = runTopology(spec, dag, { data, context })
+    const { promise, getSnapshot } = runTopology(spec, { data, context })
     await promise
     expect(getSnapshot()).toMatchObject({
       status: 'completed',
-      dag: { api: { deps: [] }, details: { deps: ['api'] } },
       data: {
         api: {
-          input: [1, 2, 3],
+          input: [[1, 2, 3]],
           status: 'completed',
           output: {
-            data: [1, 2, 3],
-            resources: { elasticCloud: 'elastic' },
+            data: [[1, 2, 3]],
             context: { launchMissleCode: 1234 },
           },
         },
         details: {
           input: [
             {
-              data: [1, 2, 3],
-              resources: { elasticCloud: 'elastic' },
+              data: [[1, 2, 3]],
               context: { launchMissleCode: 1234 },
             },
           ],
@@ -376,125 +314,72 @@ describe('runTopology', () => {
           output: {
             data: [
               {
-                data: [1, 2, 3],
-                resources: { elasticCloud: 'elastic' },
+                data: [[1, 2, 3]],
                 context: { launchMissleCode: 1234 },
               },
             ],
-            resources: { mongoDb: 'mongo' },
             context: { launchMissleCode: 1234 },
           },
         },
       },
     })
   })
-  test('bad arguments', () => {
-    const dag = {
-      api: { deps: [] },
-      details: { deps: ['api'] },
-      attachments: { deps: ['api'] },
-      writeToDB: { deps: ['details', 'attachments'] },
-    }
-    const spec: Spec = {
-      nodes: {
-        api: {
-          run: async () => [1, 2, 3],
-        },
-        attachments: {
-          run: async () => 2,
-        },
-      },
-    }
-    expect(() => {
-      runTopology(spec, dag)
-    }).toThrow(
-      new TopologyError(
-        'Missing the following nodes in spec: details, writeToDB'
-      )
-    )
-  })
   test('completed', async () => {
     const { promise, getSnapshot } = runTopology(spec, dag)
     await promise
     expect(getSnapshot()).toMatchObject({
       status: 'completed',
-      dag: {
-        api: { deps: [] },
-        details: { deps: ['api'] },
-        attachments: { deps: ['api'] },
-        writeToDB: { deps: ['details', 'attachments'] },
-      },
       data: {
         api: {
-          input: [],
+          deps: [],
           status: 'completed',
+          input: [],
           output: [1, 2, 3],
         },
         details: {
-          input: [[1, 2, 3]],
+          deps: ['api'],
           status: 'completed',
+          input: [[1, 2, 3]],
           output: {
-            1: 'description 1',
-            2: 'description 2',
-            3: 'description 3',
+            '1': 'description 1',
+            '2': 'description 2',
+            '3': 'description 3',
           },
         },
         attachments: {
-          input: [[1, 2, 3]],
+          deps: ['api'],
           status: 'completed',
-          output: {
-            1: 'file1.jpg',
-            2: 'file2.jpg',
-            3: 'file3.jpg',
-          },
+          input: [[1, 2, 3]],
+          output: { '1': 'file1.jpg', '2': 'file2.jpg', '3': 'file3.jpg' },
         },
         writeToDB: {
+          deps: ['details', 'attachments'],
+          status: 'completed',
           input: [
             {
-              1: 'description 1',
-              2: 'description 2',
-              3: 'description 3',
+              '1': 'description 1',
+              '2': 'description 2',
+              '3': 'description 3',
             },
-            {
-              1: 'file1.jpg',
-              2: 'file2.jpg',
-              3: 'file3.jpg',
-            },
+            { '1': 'file1.jpg', '2': 'file2.jpg', '3': 'file3.jpg' },
           ],
-          status: 'completed',
           output: null,
         },
       },
     })
   })
   test('gracefully shutdown when stop is called', async () => {
-    const cleanedUp: any[] = []
-    const cleanup = (x: any) => {
-      cleanedUp.push(x)
-    }
-
-    const dag: DAG = {
-      api: { deps: [] },
-    }
     const spec: Spec = {
-      resources: {
-        elasticCloud: {
-          init: async () => 'elastic',
-          cleanup: async (x: any) => cleanup(x),
-        },
-      },
-      nodes: {
-        api: {
-          run: async ({ signal, updateState }) => {
-            for (let i = 0; i < 5; i++) {
-              if (signal.aborted) {
-                throw new Error('Aborted')
-              }
-              await timers.setTimeout(100)
-              updateState({ index: i })
+      api: {
+        deps: [],
+        run: async ({ signal, updateState }) => {
+          for (let i = 0; i < 5; i++) {
+            if (signal.aborted) {
+              throw new Error('Aborted')
             }
-          },
-          resources: ['elasticCloud'],
+            await timers.setTimeout(100)
+            updateState({ index: i })
+          }
         },
       },
     }
@@ -505,18 +390,15 @@ describe('runTopology', () => {
     // Node errored
     expect(getSnapshot()).toMatchObject({
       status: 'errored',
-      dag: { api: { deps: [] } },
       data: {
         api: {
           input: [],
           status: 'errored',
           state: { index: 1 },
-          error: 'Error: Aborted',
+          error: { stack: expect.stringContaining('Error: Aborted') },
         },
       },
     })
-    // Resources cleaned up
-    expect(cleanedUp).toEqual(['elastic'])
   })
 })
 
@@ -526,14 +408,9 @@ describe('getResumeSnapshot', () => {
       status: 'errored',
       started: new Date('2020-01-01T00:00:00Z'),
       finished: new Date('2020-01-01T00:00:01Z'),
-      dag: {
-        api: { deps: [] },
-        details: { deps: ['api'] },
-        attachments: { deps: ['api'] },
-        writeToDB: { deps: ['details', 'attachments'] },
-      },
       data: {
         api: {
+          deps: [],
           started: new Date('2020-01-01T00:00:00Z'),
           finished: new Date('2020-01-01T00:00:01Z'),
           input: [],
@@ -541,6 +418,7 @@ describe('getResumeSnapshot', () => {
           output: [1, 2, 3],
         },
         details: {
+          deps: ['api'],
           started: new Date('2020-01-01T00:00:00Z'),
           finished: new Date('2020-01-01T00:00:01Z'),
           input: [[1, 2, 3]],
@@ -552,6 +430,7 @@ describe('getResumeSnapshot', () => {
           },
         },
         attachments: {
+          deps: ['api'],
           started: new Date('2020-01-01T00:00:00Z'),
           input: [[1, 2, 3]],
           status: 'errored',
@@ -569,12 +448,6 @@ describe('getResumeSnapshot', () => {
     expect(snapshot.finished).toBeUndefined()
     expect(snapshot).toMatchObject({
       status: 'running',
-      dag: {
-        api: { deps: [] },
-        details: { deps: ['api'] },
-        attachments: { deps: ['api'] },
-        writeToDB: { deps: ['details', 'attachments'] },
-      },
       data: {
         api: {
           started: new Date('2020-01-01T00:00:00.000Z'),
@@ -625,7 +498,7 @@ describe('resumeTopology', () => {
     }
     return output
   }
-  const modifiedSpec = _.set('nodes.attachments.run', attachmentsRun, spec)
+  const modifiedSpec = _.set('attachments.run', attachmentsRun, spec)
 
   test('resume after initial error', async () => {
     const { promise, getSnapshot } = runTopology(modifiedSpec, dag)
@@ -633,19 +506,15 @@ describe('resumeTopology', () => {
     const snapshot = getSnapshot()
     expect(snapshot).toMatchObject({
       status: 'errored',
-      dag: {
-        api: { deps: [] },
-        details: { deps: ['api'] },
-        attachments: { deps: ['api'] },
-        writeToDB: { deps: ['details', 'attachments'] },
-      },
       data: {
         api: {
+          deps: [],
           input: [],
           status: 'completed',
           output: [1, 2, 3],
         },
         details: {
+          deps: ['api'],
           input: [[1, 2, 3]],
           status: 'completed',
           output: {
@@ -655,9 +524,12 @@ describe('resumeTopology', () => {
           },
         },
         attachments: {
+          deps: ['api'],
           input: [[1, 2, 3]],
           status: 'errored',
-          error: 'Error: Failed processing id: 2',
+          error: {
+            stack: expect.stringContaining('Error: Failed processing id: 2'),
+          },
           state: {
             index: 0,
             output: {
@@ -672,19 +544,15 @@ describe('resumeTopology', () => {
     await promise2
     expect(getSnapshot2()).toMatchObject({
       status: 'completed',
-      dag: {
-        api: { deps: [] },
-        details: { deps: ['api'] },
-        attachments: { deps: ['api'] },
-        writeToDB: { deps: ['details', 'attachments'] },
-      },
       data: {
         api: {
+          deps: [],
           input: [],
           status: 'completed',
           output: [1, 2, 3],
         },
         details: {
+          deps: ['api'],
           input: [[1, 2, 3]],
           status: 'completed',
           output: {
@@ -694,6 +562,7 @@ describe('resumeTopology', () => {
           },
         },
         attachments: {
+          deps: ['api'],
           input: [[1, 2, 3]],
           status: 'completed',
           output: {
@@ -703,6 +572,7 @@ describe('resumeTopology', () => {
           },
         },
         writeToDB: {
+          deps: ['details', 'attachments'],
           input: [
             {
               1: 'description 1',
@@ -726,21 +596,20 @@ describe('resumeTopology', () => {
       started: new Date('2022-01-01T12:00:00Z'),
       finished: new Date('2022-01-01T12:00:01Z'),
       status: 'completed',
-      dag: { api: { deps: [] }, details: { deps: ['api'] } },
       data: {
         api: {
+          deps: [],
           input: [1, 2, 3],
           status: 'completed',
           output: {
             data: [1, 2, 3],
-            resources: { elasticCloud: 'elastic' },
           },
         },
         details: {
+          deps: ['api'],
           input: [
             {
               data: [1, 2, 3],
-              resources: { elasticCloud: 'elastic' },
             },
           ],
           status: 'completed',
@@ -748,10 +617,8 @@ describe('resumeTopology', () => {
             data: [
               {
                 data: [1, 2, 3],
-                resources: { elasticCloud: 'elastic' },
               },
             ],
-            resources: { mongoDb: 'mongo' },
           },
         },
       },