+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.7.0
+
+* Spec includes `deps` so you can just pass one object.
+* Snapshot data includes `deps`. Removed `dag`.
+* Get error stacktrace if available.
+* Wrap initial input data in an array to match other nodes.
+* Removed `resources`.
+
 # 0.6.0
 
 * Handle empty snapshot when resuming.